Struct guppy_grpc::grpc_client::Request

pub struct Request<T> { /* private fields */ }

A gRPC request and metadata from an RPC call.

Implementations

impl<T> Request<T>

pub fn new(message: T) -> Request<T>

Create a new gRPC request.

Request::new(HelloRequest {
   name: "Bob".into(),
});

pub fn get_ref(&self) -> &T

Get a reference to the message

pub fn get_mut(&mut self) -> &mut T

Get a mutable reference to the message

pub fn metadata(&self) -> &MetadataMap

Get a reference to the custom request metadata.

pub fn metadata_mut(&mut self) -> &mut MetadataMap

Get a mutable reference to the request metadata.

pub fn into_inner(self) -> T

Consumes self, returning the message

pub fn from_http(http: Request<T>) -> Request<T>

Convert an HTTP request to a gRPC request

pub fn remote_addr(&self) -> Option<SocketAddr>

Get the remote address of this connection.

This will return None if the IO type used does not implement Connected. This currently, only works on the server side.

pub fn peer_certs(&self) -> Option<Arc<Vec<Certificate, Global>, Global>>

Get the peer certificates of the connected client.

This is used to fetch the certificates from the TLS session and is mostly used for mTLS. This currently only returns Some on the server side of the transport server with TLS enabled connections.